requirements.txt
pandas
openpyxl
gradio
python-dotenv
google-generativeai
